requirements.txt
transformers
torch
Pillow
streamlit
gradio
